% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_vb_mle.R
\name{predict_single_model}
\alias{predict_single_model}
\title{Helper function for prediction}
\usage{
predict_single_model(model, newdata, interval, level)
}
\arguments{
\item{model}{A single vb_optim model}

\item{newdata}{Data frame with ages}

\item{interval}{Type of interval}

\item{level}{Confidence level}
}
\value{
Data frame with predictions
}
\description{
Helper function for prediction
}
\keyword{internal}
