% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lw.R
\name{print.lw_fit}
\alias{print.lw_fit}
\title{Print Method for Length-Weight Fit Objects}
\usage{
\method{print}{lw_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lw_fit}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the object
}
\description{
Print Method for Length-Weight Fit Objects
}
