% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_vb_mle.R
\name{print.vb_mle}
\alias{print.vb_mle}
\title{Print method for vb_mle objects}
\usage{
\method{print}{vb_mle}(x, digits = 4, ...)
}
\arguments{
\item{x}{A fitted model object of class vb_mle}

\item{digits}{Number of digits to display (default 4)}

\item{...}{Additional arguments (not used)}
}
\value{
The object x, invisibly
}
\description{
Print method for vb_mle objects
}
