% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_vb_mle.R
\name{print_single_model}
\alias{print_single_model}
\title{Helper function for printing model results}
\usage{
print_single_model(model, digits)
}
\arguments{
\item{model}{A single vb_optim model}

\item{digits}{Number of digits to display}
}
\description{
Helper function for printing model results
}
\keyword{internal}
