% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lw.R
\name{summary.lw_fit}
\alias{summary.lw_fit}
\title{Summary Method for Length-Weight Fit Objects}
\usage{
\method{summary}{lw_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lw_fit}}

\item{...}{Additional arguments (not used)}
}
\value{
A list containing summary information
}
\description{
Summary Method for Length-Weight Fit Objects
}
