% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_vb_mle.R
\name{summary.vb_mle}
\alias{summary.vb_mle}
\title{Summary method for vb_mle objects}
\usage{
\method{summary}{vb_mle}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class vb_mle}

\item{...}{Additional arguments (not used)}
}
\value{
A summary object
}
\description{
Summary method for vb_mle objects
}
