% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_vb_mle.R
\name{summary_single_model}
\alias{summary_single_model}
\title{Helper function for summarising a single model}
\usage{
summary_single_model(model)
}
\arguments{
\item{model}{A single vb_optim model}
}
\description{
Helper function for summarising a single model
}
\keyword{internal}
